Rename LongDoubleCoverterUtil
It's to allow adding appropriate util methods for FLOAT in next commit.
